+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 ## [0.3.1] - Unreleased
+### Fixed
+- URL redirection bug (converted `target="_blank"` into `openUrl()`)
+- Bug from Accessing Custom and File Scans directly from the overview page.
+### Changed
+- Expanded the Credits Section with more credits
 
 ## [0.3.0] - 2026-01-25
 ### Added